Separate Last.FM client and scrobbler
