+++ scrobble/lastfm/client_test.go
